+++ src/ReadonlyTuple.ts
@@ -5,9 +5,8 @@ import { Applicative, Applicative2C } from './Applicative'
 import { Apply2C } from './Apply'
 import { Bifunctor2 } from './Bifunctor'
 import { Chain2C } from './Chain'
-import { ChainRec2C } from './ChainRec'
 import { Comonad2 } from './Comonad'
-import { Either } from './Either'
+import { Extend2 } from './Extend'
 import { Foldable2 } from './Foldable'
 import { identity } from './function'
 import { Functor2 } from './Functor'
@@ -17,7 +16,6 @@ import { Monoid } from './Monoid'
 import { Semigroup } from './Semigroup'
 import { Semigroupoid2 } from './Semigroupoid'
 import { PipeableTraverse2, Traversable2 } from './Traversable'
-import { Extend2 } from './Extend'
 
 // -------------------------------------------------------------------------------------
 // model
@@ -110,35 +108,6 @@ export function getMonad<S>(M: Monoid<S>): Monad2C<URI, S> {
     ap: C.ap,
     chain: C.chain,
     of: of(M)
-  }
-}
-
-// TODO: remove in v3
-/**
- * @category instances
- * @since 2.5.0
- */
-export function getChainRec<S>(M: Monoid<S>): ChainRec2C<URI, S> {
-  const chainRec = <A, B>(a: A, f: (a: A) => readonly [Either<A, B>, S]): readonly [B, S] => {
-    let result: readonly [Either<A, B>, S] = f(a)
-    let acc: S = M.empty
-    let s: Either<A, B> = fst(result)
-    while (s._tag === 'Left') {
-      acc = M.concat(acc, snd(result))
-      result = f(s.left)
-      s = fst(result)
-    }
-    return [s.right, M.concat(acc, snd(result))]
-  }
-
-  const C = getChain(M)
-  return {
-    URI,
-    _E: undefined as any,
-    map: C.map,
-    ap: C.ap,
-    chain: C.chain,
-    chainRec
   }
 }