+++ src/index.ts
@@ -17,7 +17,6 @@ import * as boundedLattice from './BoundedLattice'
 import * as boundedMeetSemilattice from './BoundedMeetSemilattice'
 import * as category from './Category'
 import * as chain from './Chain'
-import * as chainRec from './ChainRec'
 import * as choice from './Choice'
 import * as comonad from './Comonad'
 import * as compactable from './Compactable'
@@ -155,10 +154,6 @@ export {
    * @since 2.0.0
    */
   chain,
-  /**
-   * @since 2.0.0
-   */
-  chainRec,
   /**
    * @since 2.0.0
    */